Gradle task to process the js annotations
