+++ html4j-maven-plugin/src/main/java/org/netbeans/html/mojo/ProcessJsAnnotationsMojo.java
@@ -54,16 +54,21 @@ import org.objectweb.asm.ClassReader;
 public final class ProcessJsAnnotationsMojo extends AbstractMojo {
     @Component
     private MavenProject prj;
-    
+
     @Parameter(defaultValue = "${project.build.directory}/classes")
     private File classes;
-    
+
     public ProcessJsAnnotationsMojo() {
     }
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
-        LinkedList<URL> arr = new LinkedList<URL>();
+        ProcessJsAnnotations jsa = new ProcessJsAnnotations() {
+            @Override
+            protected void log(String msg) {
+                getLog().info(msg);
+            }
+        };
         boolean foundAsm = false;
         for (Artifact a : prj.getArtifacts()) {
             final File f = a.getFile();
@@ -71,129 +76,16 @@ public final class ProcessJsAnnotationsMojo extends AbstractMojo {
                 if (a.getArtifactId().equals("asm")) {
                     foundAsm = true;
                 }
-                try {
-                    arr.add(f.toURI().toURL());
-                } catch (MalformedURLException ex) {
-                    throw new IllegalStateException(ex);
-                }
+                jsa.addClasspathEntry(f);
             }
         }
         if (!foundAsm) {
-            URL loc = ClassReader.class.getProtectionDomain().getCodeSource().getLocation();
-            arr.addFirst(loc);
+            jsa.addAsm();
         }
         try {
-            arr.add(classes.toURI().toURL());
-            URLClassLoader l = new URLClassLoader(arr.toArray(new URL[arr.size()]));
-            File master = new File(new File(classes, "META-INF"), "net.java.html.js.classes");
-            processClasses(l, master, classes);
+            jsa.process(classes);
         } catch (IOException ex) {
             throw new MojoExecutionException("Problem converting JavaScriptXXX annotations", ex);
-        }
-    }
-    
-    private void processClasses(ClassLoader l, File master, File f) throws IOException, MojoExecutionException {
-        if (!f.exists()) {
-            return;
-        }
-        if (f.isDirectory()) {
-            boolean classes = new File(f, "net.java.html.js.classes").exists();
-            File[] arr = f.listFiles();
-            if (arr != null) {
-                for (File file : arr) {
-                    if (classes || file.isDirectory()) {
-                        processClasses(l, master, file);
-                    }
-                }
-            }
-            return;
-        }
-        
-        if (!f.getName().endsWith(".class")) {
-            return;
-        }
-        
-        byte[] arr = new byte[(int)f.length()];
-        FileInputStream is = new FileInputStream(f);
-        try {
-            readArr(arr, is);
-        } finally {
-            is.close();
-        }
-
-        byte[] newArr = null;
-        try {
-            Class<?> fnUtils = l.loadClass("org.netbeans.html.boot.impl.FnUtils");
-            Method transform = fnUtils.getMethod("transform", byte[].class, ClassLoader.class);
-            
-            newArr = (byte[]) transform.invoke(null, arr, l);
-            if (newArr == null || newArr == arr) {
-                return;
-            }
-            filterClass(new File(f.getParentFile(), "net.java.html.js.classes"), f.getName());
-            filterClass(master, f.getName());
-        } catch (Exception ex) {
-            throw new MojoExecutionException("Can't process " + f, ex);
-        }
-        getLog().info("Processing " + f);
-        writeArr(f, newArr);        
-    }
-
-    private void writeArr(File f, byte[] newArr) throws IOException, FileNotFoundException {
-        FileOutputStream os = new FileOutputStream(f);
-        try {
-            os.write(newArr);
-        } finally {
-            os.close();
-        }
-    }
-
-    private static void readArr(byte[] arr, InputStream is) throws IOException {
-        int off = 0;
-        while (off< arr.length) {
-            int read = is.read(arr, off, arr.length - off);
-            if (read == -1) {
-                break;
-            }
-            off += read;
-        }
-    }
-    
-    private static void filterClass(File f, String className) throws IOException {
-        if (!f.exists()) {
-            return;
-        }
-        if (className.endsWith(".class")) {
-            className = className.substring(0, className.length() - 6);
-        }
-        
-        BufferedReader r = new BufferedReader(new FileReader(f));
-        List<String> arr = new ArrayList<String>();
-        boolean modified = false;
-        for (;;) {
-            String line = r.readLine();
-            if (line == null) {
-                break;
-            }
-            if (line.endsWith(className)) {
-                modified = true;
-                continue;
-            }
-            arr.add(line);
-        }
-        r.close();
-        
-        if (modified) {
-            if (arr.isEmpty()) {
-                f.delete();
-            } else {
-                FileWriter w = new FileWriter(f);
-                for (String l : arr) {
-                    w.write(l);
-                    w.write("\n");
-                }
-                w.close();
-            }
         }
     }
 }
